pyproject.toml
[project]
name = "n8n-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "langchain-core>=1.3.0",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.8",
    "openpyxl>=3.1.5",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]
